Remove obsolete code paths
