+++ agent/helpudoc_agent/api/__init__.py
@@ -1,0 +1,1 @@
+"""Agent HTTP API package (routes, schemas, app factory)."""
