+++ agent/helpudoc_agent/api/routes/__init__.py
@@ -1,0 +1,1 @@
+"""Route registration subpackage."""
