+++ agent/main.py
@@ -5,6 +5,6 @@ CURRENT_DIR = Path(__file__).resolve().parent
 if str(CURRENT_DIR) not in sys.path:
     sys.path.insert(0, str(CURRENT_DIR))
 
-from helpudoc_agent.app import create_app
+from helpudoc_agent.api.app import create_app
 
 app = create_app()
